+++ app/cnn/model_2/variables/variables.data-00000-of-00001
