+++ test.py
@@ -1,10 +1,10 @@
 import unittest
 import numpy as np
 
-from app.depthFirstSearch.depthFirstSearch import depth_first_search
-from app.depthFirstSearch import stateFactory, state
-from app.depthFirstSearch.state import State
-from app.neuralNetwork.convNet import one_hot_encode, decode_output
+from app.dfs.depthFirstSearch import depth_first_search
+from app.dfs import stateFactory, state
+from app.dfs.state import State
+from app.cnn.convNet import one_hot_encode, decode_output
 from app.utilites import string_to_np_array, np_array_to_string
